requirements.txt
streamlit
pandas
pyarrow
reportlab
mysqlconnectorpython
pythondotenv
pillow
duckdb